Proof of concept
So far C++ part of project is not included in this repo. The C++ code was developed inside the SDK directory folder by modifying an example. Once I have got the hang of how to set up projects in Visual Studio it is planned to be moved inside the repo.