Use new exe
Trying out the exe built within this repo. Should be identical to the previous one, except for the name which is now in camelcase.

Stop Git from ignoring bin. Ruby developers should be able to update the Ruby code in this project without having to set up VS and re-compile the external C++ program.